Drop @skip in test-aa-notify
The Perl aa-notify script was written so that it will check for kern.log
before printing help when invoked without arguments (sic!). Therefore,
the test for calling aa-notify witout a parameter was skipped if
/var/log/kern.log doesn't exist.

The python version does not need kern.log to print the help, therefore
drop the @skip.

After that, the test becomes very similar to the test for
"aa-notify --help". De-duplicte the code so that both "aa-notify" and
"aa-notify --help" tests use the same code.