parser: allow reducing the restrictiveness of mount option handling
AppArmor 4.0 as part of the fix for mount rule generation in
commit ecfaf7330 ("parser: add conflicting flags check for options=
  conditionals")

added a restrictive check on the mount flags that could be passed to an
exact match option. This allowed mount options that can not be set at the
same time like 'ro' and 'rw' to be caught and reported as a semantic
failure.
  Eg.
  options=(rw ro)

Unfortunately rules doing this exist in existing policy. So add an
option to allow the conflict in rules, by generating an option flag
set that allows both. Effectively options=(rw) options=(ro).

This relaxation of the conflict check is enabled by setting the
parser option
  --allow-mount-conflict

Signed-off-by: John Johansen <john.johansen@canonical.com>