Delete copy constructors in parser classes that have pointers as per the rule of 3
Move constructors are not included because they don't seem to be used and
we can write them if needed later.

Another option would be to use unique_ptr instead of raw pointers, which
would be a good direction to move in long-term, but is also a bigger
change than I feel comfortable making all at once with my current level of
understanding of the parser codebase.

Signed-off-by: Ryan Lee <ryan.lee@canonical.com>