GitLab CI: Add Alpine CI to test musl libc
Testing under Alpine allows to test against a different libc as glibc
(Ubuntu) which can find build failures on non-glibc before release.

It is configured to allow to fail, so that it is not a blocker for MRs
or releases.

I would volunteer to maintain the Alpine CI and for any build failures,
feel free to ping me and I can try to solve them.