tests: run spread integration tests
This requires a runner with the tags: linux, x86_64, kvm. One needs to
be provisioned for the AppArmor project for the pipeline to function.

It is possible to run the same tests on SAAS runners offered by GitLab
but due to issue https://gitlab.com/gitlab-org/gitlab-runner/-/issues/6208
there is no way to expose /dev/kvm on the host to the guest. Without this
feature emulation works but is rather slow as to be impractical.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>