tests: use one worker per backend
This is somewhat unfortunate, but we are unable to both have wide local testing
and wide testing (by sharding) in CI/CD pipelines as the place where widening
happens needs to be different.

For the moment prefer wide testing in CI/CD until this is improved in spread
itself, with a command line switch that overrides the number of declared workers.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>