utils/test/test-aa-notify.py: update last cmd for lxd VMs
When doing testing via LXD VMs and in particular when using "lxc exec" to run
commands in the VM, there is no controlling tty and so the output of last is
missing this column of data. Instead try even harder to parse the timestamp from
the output of "last".

Signed-off-by: Alex Murray <alex.murray@canonical.com>
(cherry picked from commit b6654416b0433f077809c7f1d2f177389b0143d1)
Signed-off-by: John Johansen <john.johansen@canonical.com>