(WIP) adds a charon-systemd profile for strongswan
A charon-systemd profile has existed for years in the strongswan source
package. It has however always been in complain mode and is not actively
maintained. This is an attempt at making an up-to-date enforceable new
profile. For now, only the most basic host-to-host configuration has
been tested. Pings work between two up-to-date ubuntu oracular machines.

Signed-off-by: Noam Nedelec-Salmon <noam.nedelecsalmon@canonical.com>